deployment_cicd_seed/pyproject.toml
[tool.poetry]
name = "mlflow-model-deployment"
version = "0.1.0"
description = ""
authors = ["Alessandro Cere <alecere@amazon.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "2.29.1"
constructs = ">=10.0.0,<11.0.0"
boto3 = "^1.24.51"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
